修复：视频上传mediaType改为VIDEO+消息计数Integer解析+弹窗定位修正+移除发送成功提示+视频压缩

+++ app/src/main/java/com/flashnote/java/data/remote/MessageService.java
@@ -24,5 +24,5 @@ public interface MessageService {
     Call<ApiResponse<Void>> delete(@Path("id") Long id);
 
     @GET("api/messages/count")
-    Call<ApiResponse<Long>> countMessages();
+    Call<ApiResponse<Integer>> countMessages();
 }
